Update MainActivity.kt
Modified the code to include a Plotter and two sliders one for X-Axis and another for Y-Axis. Changes in the Slider results in the plotting point to move on the plotter surface.

Constants like size of Outer box and inner box and offset value of -18 is defined as variables in the Column Composable.

Function Map() is not using xPercent and yPercent, function Map() was already declared in the starter project with these parameters and modifier. So I created modifier variable while calling the function itself.

Modified the inner-box size to 20 to make it look better.

Horizontally aligned the Plotter Box to the Center of the Column.